refactor: 직관내역 리스트 화면 컴포저블 분리 및 상태 관리 로직 개선
